+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Rely on `OpenSSL::SSL.verify_certificate_identity` to check that a certificate
+  is valid for the provided hostname.
+
 ## [1.1.0] - 2019-02-21
+
 ### Added
 - Add basic suport for TLSA events.